requirements.txt
onnx
onnxruntime
opencv-python
numpy
scikit-image
faiss-cpu
flask
werkzeug
pymilvus
boto3
bleach
python-dotenv
scikit-learn
gunicorn
gevent